Signal: retain the sample_rate on create_new
When using "Create signal from selection" feature, the newly created
signal does not retain sample rate of the original so all the time of
the selection is computed wrong which is very confusing to the user.

This simple patch makes sure the sample_rate of the original Signal is
copied to the newly created one which fixes the problem.