Fix .sub file RAW_data strip
For some reason some of the files saved by flipper has a final space. When the RAW_data is splitted, the last space gets it's own '' value in the array. Whet we try to convert the '' value to int we get an exception.
The exception has been controlled, showing a warning and skipping the value.
The line is stripped to remove white spaces before of splitting